requirements.txt
django==1.9.4
selenium==2.53.1
fabric==1.10.2
djangorestframework==3.3.3
